feat: query consistency (Phase 6) + documentation (Phase 8)
Phase 6: Advanced API Features
- Added multi-column search to Programs (name+code), Courses
  (code+title), Academic Terms (year+semester, previously had none)
- Added search-by-related-course to Course Offerings
- Added sorting to Enrollments index (enrollment_date, created_at),
  previously filter-only
- Capped per_page at 100 across all six list/nested-list endpoints
  (Programs, Courses, Academic Terms, Students, Course Offerings +
  its /students sub-route, Enrollments + its /students and /grades
  sub-routes) to prevent unbounded dataset retrieval
- Added exception rendering for InvalidFilterQuery/InvalidSortQuery
  so bad query params return consistent 400 JSON instead of a raw
  framework error

Phase 8: Documentation
- README.md: tech stack, setup, env config, migration/seed
  instructions, demo account credentials, auth flow, testing,
  AI-assisted development summary
- docs/ERD.md: Mermaid entity relationship diagram covering all
  8 core tables, relationships, and unique constraints
- docs/endpoint-reference.md: full manual reference for every
  implemented endpoint (method, auth/role, example body,
  expected success/failure responses)
- Confirmed Scramble OpenAPI docs render at /docs/api

Known gap (deferred): Course Offerings has no role/policy
restriction yet — any authenticated user can currently
create/edit/delete offerings. To be addressed in Phase 7.

+++ app/Http/Controllers/Api/V1/GradeController.php
@@ -44,11 +44,12 @@ class GradeController extends Controller
 
     public function indexForStudent(\App\Models\Student $student)
     {
-        $this->authorize('view', $student); // same rule — student sees own, admin/registrar see all, instructor blocked unless extended
-        $perPage = min(request('per_page', 20), 100); // Limit the maximum per page to 100
+        $this->authorize('view', $student);
+
+        $perPage = min((int) request('per_page', 20), 100);
+
         $grades = \App\Models\Grade::whereHas('enrollment', fn ($q) => $q->where('student_id', $student->id))
             ->with('enrollment.courseOffering.course')
-
             ->paginate($perPage);
 
         return $this->success(GradeResource::collection($grades)->response()->getData(true), 'Student grades retrieved successfully.');
